+++ R/viridis_quo.R
@@ -140,3 +140,43 @@ geom_hist_bw <- function(mapping = NULL,
     show.legend = show.legend, inherit.aes = inherit.aes
   )
 }
+
+#' Annotate using quo theme font
+#'
+#' A wrapper for [ggplot2::annotate()] that sets the font to Lato.
+#'
+#' @inheritParams ggplot2::annotate
+#' @param family Font family to use. Defaults to `"Lato"`.
+#'
+#' @examples
+#' library(ggplot2)
+#' library(showtext)
+#'
+#' font_add_google("Lato", "Lato")
+#' showtext_auto()
+#' # workaround for https://github.com/yixuan/showtext/issues/51
+#' showtext_opts(dpi = 192)
+#'
+#' ggplot(mpg, aes(displ, hwy)) +
+#'   geom_point() +
+#'   geom_rlm(method = "loess") +
+#'   annotate_quo("label", x = 5.5, y = 30, label = "hwy increases for displ greater than 5.5") +
+#'   theme_quo()
+#' @export
+annotate_quo <- function(geom,
+                         x = NULL,
+                         y = NULL,
+                         xmin = NULL,
+                         xmax = NULL,
+                         ymin = NULL,
+                         ymax = NULL,
+                         xend = NULL,
+                         yend = NULL,
+                         family = "Lato",
+                         ...,
+                         na.rm = FALSE) { # nolint: object_name_linter.
+  ggplot2::annotate(
+    geom = geom, x = x, y = y, xmin = xmin, xmax = xmax, ymin = ymin, ymax = ymax, xend = xend,
+    yend = yend, family = family, ..., na.rm = na.rm
+  )
+}